feat(ios): add Splice, StabilitySend, and PendingOperation repositories for improved database management
